requirements.txt
fastapi
uvicorn
SQLAlchemy
databases
asyncpg
passlib[bcrypt]
python-jose
pydantic
alembic
psycopg2
fastapi_utils
typing_inspect
pydantic_settings
humanize
fastapi_utils
python-barcode
bs4
pydantic[email]
python-multipart
pandas
google-auth
google-api-python-client
httpx
openpyxl
apscheduler
